+++ Sources/FlinkuSDK/FlinkuError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum FlinkuError: Error {
+    case missingApiKey
+    case invalidResponse
+    case invalidURL
+    case notConfigured
+}